requirements.txt
pandas 
numpy 
scikit-learn
pymongo
certifi
python-dotenv

-e .